etodo_api/pyproject.toml
[tool.poetry]
name = "etodo-api"
version = "0.1.0"
description = ""
authors = ["Jai Jain <jai.jain@acsicorp.com>"]
readme = "README.md"
packages = [{include = "etodo_api"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.3"
djangorestframework = "^3.14.0"
django-filter = "^23.2"
django-celery-beat = "^2.5.0"
werkzeug = "^2.3.6"
django-extensions = "^3.2.3"
drf-spectacular = "^0.26.3"
celery = {extras = ["redis"], version = "^5.3.1"}
django-guardian = "^2.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
